requirements.txt
fastapi==0.115.5
Pillow==11.0.0
ultralytics==8.3.31
uvicorn==0.32.0
